refactor: Update user profile page to display likes count
The code changes in the `UsersController.php` and `show.blade.php` files update the user profile page to display the count of likes the user has received. The `getUserLikesCount` function is added to the `UsersController` to retrieve the likes count for a user. The likes count is then passed to the `show.blade.php` view and displayed next to the thumbs-up icon.