refactor: Add city field to User model and settings page
The code changes in the User model, UserFactory, and database migration files add a new 'city' field to the User model. The city field is now included in the user registration process and can be updated in the user settings page. This allows users to specify their city information and view it on their profile page.